Callbacks for changing incoming data and validated RTP packets
Added OnValidatedRTPPacket to RTPSession and RTPSources, allowing you
to process a packet as soon as it is linked to an RTPSourceData
instance. This way you can avoid the GotoFirst/GotoNext iteration over
the sources.

An OnChangeIncomingData member function was added to RTPSession,
allowing you to change received RTP or RTCP data. This can be used to
e.g. decrypt the packets.