Added OnValidatedRTPPacket to discard (the unused) incoming RTP data
When RTPDEBUG was enabled, and memory was being tracked, not removing
the received RTP data caused a problem in a stess test: RTP packets
were continuously being stored in the RTPSources instance, and when
finally deleting this instance, the memory tracker had become too slow
due to the amount of packets. This caused a thread to be forcefully
killed, in turn causing a mutex to hang in the memory tracker. That was
the cause of the program not getting shut down cleanly.