Initial version of a TCP transmitter
To avoid having to do the connection setup in the rtp library,
you control which connections are used for sending and receiving
by adding sockets (using RTPTCPAddress wrapper) in AddDestination
calls. In this case, the name 'AddDestination' is not very accurate
as the same socket will also be used to check for incoming data.